organise in a similar fashion to react-redux example
